refactor blogController

+++ src/routes/blogs.ts
@@ -2,7 +2,14 @@ import express from 'express';
 import type { Request, Response } from 'express';
 import { authMiddleware } from '../middleware/auth';
 import { validateBlog } from '../middleware/validation';
-import { createBlog, deleteBlog, getBlogs, updateBlog } from '../controllers/blogController';
+import {
+  checkOwnership,
+  createBlog,
+  deleteBlog,
+  getBlogById,
+  getBlogs,
+  updateBlog,
+} from '../controllers/blogController';
 // Ensure custom Express types are loaded
 /// <reference path="../../types/express.d.ts" />
 
@@ -16,25 +23,38 @@ router.post(
     console.log('req.user: CREATE Blog', (req as any).user);
     // create new blog
     createBlog(req, res);
-    return ;
+    return;
   }
 );
 // get all blog posts
-router.get('/', [authMiddleware ], async (req: Request, res: Response) => {
- getBlogs(req, res);
+router.get('/', authMiddleware, async (req: Request, res: Response) => {
+  getBlogs(req, res);
 });
 // update blog post by id
-router.put('/:id', [authMiddleware , validateBlog], async (req: Request, res: Response) => {
-  // update blog 
+router.put('/:id', authMiddleware, async (req: Request, res: Response) => {
+  // check if request body is empty
+  if (Object.keys(req.body).length === 0)
+    return res.status(400).send('No data provided to update');
+  // get blog
+  const blog = await getBlogById(req, res);
+  // check ownership
+  if (checkOwnership(blog, (req as any).user))
+    return res.status(400).send('you are not allowed to update this blog');
+  // update blog
   updateBlog(req, res);
-  return ;
+  return;
 });
 
 // delete blog post by id
 router.delete('/:id', [authMiddleware], async (req: Request, res: Response) => {
+  // get blog
+  const blog = await getBlogById(req, res);
+  // check ownership
+  if (checkOwnership(blog, (req as any).user))
+    return res.status(400).send('you are not allowed to delete this blog');
   // delete blog
   deleteBlog(req, res);
-  return ;
+  return;
 });
 
 export default router;
